python/application/requirements.txt
# Program requirements
adafruit-blinka
adafruit-circuitpython-mpr121
python-socketio[client]
python-engineio
websocket-client
requests